requirements/requirements.dev.txt
fastapi
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-jose
passlib
uvicorn
python-multipart
python-dotenv
pydantic-settings